% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg.mtx.R
\docType{data}
\name{aggreg.mtx}
\alias{aggreg.mtx}
\title{Drosophila Melanogaster differential agggregated pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as regions feature.}
\format{
A list of object of matrix.
}
\usage{
data(aggreg.mtx)
}
\description{
Drosophila Melanogaster differential agggregated pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as regions feature.
}
\examples{
data(aggreg.mtx)
aggreg.mtx
}
\keyword{datasets}
