% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors_Index.gnr.R
\docType{data}
\name{anchors_Index.gnr}
\alias{anchors_Index.gnr}
\title{Drosophila Melanogaster Beaf32 indexed peaks on 2L and 2R chromosomes.}
\format{
An object of class GRanges.
}
\usage{
data(anchors_Index.gnr)
}
\description{
Drosophila Melanogaster Beaf32 indexed peaks on 2L and 2R chromosomes.
}
\examples{
data(anchors_Index.gnr)
anchors_Index.gnr
}
\keyword{datasets}
