% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors_Peaks.gnr.R
\docType{data}
\name{anchors_Peaks.gnr}
\alias{anchors_Peaks.gnr}
\title{Drosophila Melanogaster Beaf32 peaks on 2L and 2R chromosomes.}
\format{
An object of class GRanges.
}
\usage{
data(anchors_Peaks.gnr)
}
\description{
Drosophila Melanogaster Beaf32 peaks on 2L and 2R chromosomes.
}
\examples{
data(anchors_Peaks.gnr)
anchors_Peaks.gnr
}
\keyword{datasets}
