% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baits_Index.gnr.R
\docType{data}
\name{baits_Index.gnr}
\alias{baits_Index.gnr}
\title{Drosophila Melanogaster indexed Transcription starting sites on 2L and 2R chromosomes.}
\format{
An object of class GRanges.
}
\usage{
data(baits_Index.gnr)
}
\description{
Data from a CHip Seq experiment
}
\examples{
data(baits_Index.gnr)
baits_Index.gnr
}
\keyword{datasets}
