% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baits_Peaks.gnr.R
\docType{data}
\name{baits_Peaks.gnr}
\alias{baits_Peaks.gnr}
\title{Drosophila Melanogaster Transcription starting sites on 2L and 2R chromosomes.}
\format{
An object of class GRanges.
}
\usage{
data(baits_Peaks.gnr)
}
\description{
Data from a CHip Seq experiment
}
\examples{
data(baits_Peaks.gnr)
baits_Peaks.gnr
}
\keyword{datasets}
