% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains_Border.gnr.R
\docType{data}
\name{domains_Border.gnr}
\alias{domains_Border.gnr}
\title{Drosophila Melanogaster Tads borders on chromosome 2R and 2L (F.Ramirez 2018, https://doi.org/10.1038/s41467-017-02525-w)}
\format{
An object of class GRanges.
}
\usage{
data(domains_Border.gnr)
}
\description{
Data from a HiC Seq experiment
}
\examples{
data(domains_Border.gnr)
domains_Border.gnr
}
\keyword{datasets}
