% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains_Border_Bin.gni.R
\docType{data}
\name{domains_Border_Bin.gni}
\alias{domains_Border_Bin.gni}
\title{Drosophila Melanogaster Tads borders indexed on chromosome 2R and 2L (F.Ramirez 2018, https://doi.org/10.1038/s41467-017-02525-w)}
\format{
An object of class GInteractions.
}
\usage{
data(domains_Border_Bin.gni)
}
\description{
Data from a HiC Seq experiment
}
\examples{
data(domains_Border_Bin.gni)
domains_Border_Bin.gni
}
\keyword{datasets}
