% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.gnr.r
\docType{data}
\name{GRanges.gnr}
\alias{GRanges.gnr}
\title{Drosophila Melanogaster Beaf32 peaks on 2L and 2R chromosomes.}
\format{
An object of class GRanges.
}
\usage{
data(GRanges.gnr)
}
\description{
Drosophila Melanogaster Beaf32 peaks on 2L and 2R chromosomes.
}
\examples{
data(GRanges.gnr)
GRanges.gnr
}
\keyword{datasets}
