% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC.cmx_lst.R
\docType{data}
\name{HiC.cmx_lst}
\alias{HiC.cmx_lst}
\title{Drosophila Melanogaster HiC data on chromosome 2R and 2L in Beaf 32 depletion.}
\format{
A a list of ContactMatrix objects. Each element correspond to the interaction matrix of two chromosomes.
}
\usage{
data(HiC.cmx_lst)
}
\description{
Data from a HiC Seq experiment
}
\examples{
data(HiC.cmx_lst)
HiC.cmx_lst
}
\keyword{datasets}
