% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC.gnp.R
\docType{data}
\name{HiC.gnp}
\alias{HiC.gnp}
\title{Drosophila Melanogaster HiC data on chromosome 2R and 2L in Beaf32 depletion.}
\format{
An object of class Pairs GRanges.
}
\usage{
data(HiC.gnp)
}
\description{
Data from a HiC Seq experiment
}
\examples{
data(HiC.gnp)
HiC.gnp
}
\keyword{datasets}
