% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_ctrl.cmx_lst.R
\docType{data}
\name{HiC_ctrl.cmx_lst}
\alias{HiC_ctrl.cmx_lst}
\title{Drosophila Melanogaster HiC data on chromosome 2R and 2L in Luc depletion.}
\format{
A a list of ContactMatrix objects. Each element correspond to the interaction matrix of two chromosomes.
}
\usage{
data(HiC_ctrl.cmx_lst)
}
\description{
Data from a HiC Seq experiment
}
\examples{
data(HiC_ctrl.cmx_lst)
HiC_ctrl.cmx_lst
}
\keyword{datasets}
