% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_ctrl.gnp.r
\docType{data}
\name{HiC_ctrl.gnp}
\alias{HiC_ctrl.gnp}
\title{Drosophila Melanogaster HiC data on chromosome 2R and 2L in Luc depletion.}
\format{
An object of class Pairs GRanges.
}
\usage{
data(HiC_ctrl.gnp)
}
\description{
Data from a HiC Seq experiment
}
\examples{
data(HiC_ctrl.gnp)
HiC_ctrl.gnp
}
\keyword{datasets}
