% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.gni.R
\docType{data}
\name{interactions.gni}
\alias{interactions.gni}
\title{Pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes.}
\format{
An object of class GInteractions.
}
\usage{
data(interactions.gni)
}
\description{
Data from a ChIp Seq experiment
}
\examples{
data(interactions.gni)
interactions.gni
}
\keyword{datasets}
