% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submatrixPF.mtx_lst.R
\docType{data}
\name{submatrixPF.mtx_lst}
\alias{submatrixPF.mtx_lst}
\title{Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as points feature.}
\format{
A list of object of matrix.
}
\usage{
data(submatrixPF.mtx_lst)
}
\description{
Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as points feature.
}
\examples{
data(submatrixPF.mtx_lst)
submatrixPF.mtx_lst
}
\keyword{datasets}
