% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submatrixPF_Ctrl.mtx_lst.R
\docType{data}
\name{submatrixPF_Ctrl.mtx_lst}
\alias{submatrixPF_Ctrl.mtx_lst}
\title{Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as points feature in control HiC.}
\format{
A list of object of matrix.
}
\usage{
data(submatrixPF_Ctrl.mtx_lst)
}
\description{
Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as points feature in control HiC.
}
\examples{
data(submatrixPF_Ctrl.mtx_lst)
submatrixPF_Ctrl.mtx_lst
}
\keyword{datasets}
