% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submatrixRF.mtx_lst.R
\docType{data}
\name{submatrixRF.mtx_lst}
\alias{submatrixRF.mtx_lst}
\title{Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as regions feature.}
\format{
A list of object of matrix.
}
\usage{
data(submatrixRF.mtx_lst)
}
\description{
Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as regions feature.
}
\examples{
data(submatrixRF.mtx_lst)
submatrixRF.mtx_lst
}
\keyword{datasets}
