% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submatrixRF_Ctrl.mtx_lst.R
\docType{data}
\name{submatrixRF_Ctrl.mtx_lst}
\alias{submatrixRF_Ctrl.mtx_lst}
\title{Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as points regions in control HiC.}
\format{
A list of object of matrix.
}
\usage{
data(submatrixRF_Ctrl.mtx_lst)
}
\description{
Drosophila Melanogaster extracted pairs between Beaf-32 Peaks and TSS on 2L and 2R chromosomes. Pairs extracted as points regions in control HiC.
}
\examples{
data(submatrixRF_Ctrl.mtx_lst)
submatrixRF_Ctrl.mtx_lst
}
\keyword{datasets}
