% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tad.gnr.r
\docType{data}
\name{Tad.gnr}
\alias{Tad.gnr}
\title{Drosophila Melanogaster Tads on chromosome 2R and 2L (F.Ramirez 2018, https://doi.org/10.1038/s41467-017-02525-w)}
\format{
An object of class GRanges.
}
\usage{
data(Tad.gnr)
}
\description{
Data from a CHip Seq experiment
}
\examples{
data(Tad.gnr)
Tad.gnr
}
\keyword{datasets}
