% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tss.gnr.r
\docType{data}
\name{Tss.gnr}
\alias{Tss.gnr}
\title{Drosophila Melanogaster Transcription starting sites on 2L and 2R chromosomes.}
\format{
An object of class GRanges.
}
\usage{
data(Tss.gnr)
}
\description{
Data from a CHip Seq experiment
}
\examples{
data(Tss.gnr)
Tss.gnr
}
\keyword{datasets}
